init making validations and corrects responses

+++ orders/api.py
@@ -25,17 +25,21 @@ def specific_order(request, id: int):
 
 @orders_router.post('/', response={200: dict})
 def create(request, order_schema: OrderPostSchema):
+    if request.user.is_anonymous:
+        return raise_detail_message("Authentication", 400, "Nao tem permissao")
     new_order = Order(**order_schema.dict())
     new_order.save()
     return raise_detail_message("Created success", 201, "Encomenda criada com sucesso")
 
 @orders_unique_router.delete('/{id}')
 def delete(request, id: int):
+    if request.user.is_anonymous:
+        return raise_detail_message("Authentication", 400, "Nao tem permissao")
     delete_order = get_object_or_404(Order, id=id)
     delete_order.delete()
     return raise_detail_message("Order deleted", 200, "Objeto deletado com sucesso")
 
-# @orders_unique_router.delete('/{id}') PUT
+# @orders_unique_router.put('/{id}')
 # def delete(request, id: int):
 #     return raise_detail_message("Order deleted", 200, "Objeto deletado com sucesso")
